Add check is admin to the catalog routes as well

+++ routes/catalog.js
@@ -1,4 +1,5 @@
 // Config Variables
+
 var session = require('express-session');
 var express = require('express');
 var router = express.Router();
@@ -27,12 +28,16 @@ router.get('/', async (req, res) => {
 });
 
 router.get('/transactions', async (req, res) => {
-    try {
-        let list = await catalog.getTransactionItems();
-        res.render('transactions/transactions', {filter: false, active: "", list: await list, title: 'Transactions', is_logged: req.session.logged, is_admin: req.session.is_admin});
-    } catch (err) {
-        console.error(err);
-        res.render('error', { error: err });
+    if (currentUserIsAdmin(req)){
+        try {
+            let list = await catalog.getTransactionItems();
+            res.render('transactions/transactions', {filter: false, active: "", list: await list, title: 'Transactions', is_logged: req.session.logged, is_admin: req.session.is_admin});
+        } catch (err) {
+            console.error(err);
+            res.render('error', { error: err });
+        }
+    } else {
+        res.render('index', { title: 'Home', is_logged: req.session.logged, is_admin: req.session.is_admin, errors: [{msg: "You are not an admin!"}]});
     }
 });
 
@@ -91,21 +96,29 @@ router.post('/searchitems', async (req, res) => {
 // Page to select which item ti unsert. Upon selecting
 // the specific item create/discriminator is rendered
 router.get('/create', function (req, res) {
-    try {
-        res.render('catalog/createItems', { title: 'Create Item', is_logged: req.session.logged, is_admin: req.session.is_admin});
-    } catch (err) {
-        console.error(err);
-        res.render('error', { error: err });
+    if (currentUserIsAdmin(req)){
+        try {
+            res.render('catalog/createItems', { title: 'Create Item', is_logged: req.session.logged, is_admin: req.session.is_admin});
+        } catch (err) {
+            console.error(err);
+            res.render('error', { error: err });
+        }
+    } else {
+        res.render('index', { title: 'Home', is_logged: req.session.logged, is_admin: req.session.is_admin, errors: [{msg: "You are not an admin!"}]});
     }
 });
 // Create a new book
 router.get('/create/:discriminator', function (req, res) {
-    try {
-        let discriminator = req.params.discriminator;
-        res.render('catalog/createItem', { discriminator, title: 'Create Item', is_logged: req.session.logged, is_admin: req.session.is_admin});
-    } catch (err) {
-        console.error(err);
-        res.render('error', { error: err });
+    if (currentUserIsAdmin(req)){
+        try {
+            let discriminator = req.params.discriminator;
+            res.render('catalog/createItem', { discriminator, title: 'Create Item', is_logged: req.session.logged, is_admin: req.session.is_admin});
+        } catch (err) {
+            console.error(err);
+            res.render('error', { error: err });
+        }
+    } else {
+        res.render('index', { title: 'Home', is_logged: req.session.logged, is_admin: req.session.is_admin, errors: [{msg: "You are not an admin!"}]});
     }
 });
 
@@ -113,14 +126,18 @@ router.get('/create/:discriminator', function (req, res) {
 // == POST Requests for Creating Items === //
 // ====================================== //
 router.post('/create/:discriminator', async (req, res) => {
-    try {
-        let results = await catalog.insertNewItem(req, req.params.discriminator);
-        let item_id = results.item_id.rows[0].currval;
-        let discriminator = req.params.discriminator;
-        res.redirect('/catalog/view/'+discriminator+'/'+item_id);
-    } catch (err) {
-        console.error(err);
-        res.render('error', { error: err });
+    if (currentUserIsAdmin(req)){
+        try {
+            let results = await catalog.insertNewItem(req, req.params.discriminator);
+            let item_id = results.item_id.rows[0].currval;
+            let discriminator = req.params.discriminator;
+            res.redirect('/catalog/view/'+discriminator+'/'+item_id);
+        } catch (err) {
+            console.error(err);
+            res.render('error', { error: err });
+        }
+    } else {
+        res.render('index', { title: 'Home', is_logged: req.session.logged, is_admin: req.session.is_admin, errors: [{msg: "You are not an admin!"}]});
     }
 });
 // localhost:3000/catalog/createitems
@@ -129,13 +146,17 @@ router.post('/create/:discriminator', async (req, res) => {
 // == GET Requests for Updating Items === //
 // ====================================== //
 router.get('/update/:discriminator/:item_id', async (req, res) => {
-    try {
-        let results = await catalog.getItemById(req.params.item_id, req.params.discriminator);
-        let discriminator = results.results[0].discriminator;
-        res.render('catalog/updateitem', { results, discriminator, title: 'Catalog', is_logged: req.session.logged, is_admin: req.session.is_admin});
-    } catch (err) {
-        console.error(err);
-        res.render('error', { error: err });
+    if (currentUserIsAdmin(req)){
+        try {
+            let results = await catalog.getItemById(req.params.item_id, req.params.discriminator);
+            let discriminator = results.results[0].discriminator;
+            res.render('catalog/updateitem', { results, discriminator, title: 'Catalog', is_logged: req.session.logged, is_admin: req.session.is_admin});
+        } catch (err) {
+            console.error(err);
+            res.render('error', { error: err });
+        }
+    } else {
+        res.render('index', { title: 'Home', is_logged: req.session.logged, is_admin: req.session.is_admin, errors: [{msg: "You are not an admin!"}]});
     }
 });
 
@@ -144,15 +165,19 @@ router.get('/update/:discriminator/:item_id', async (req, res) => {
 // == POST Requests for Updating Items === //
 // ======================================= //
 router.post('/update/:discriminator/:item_id', async (req, res) => {
-    try {
-        let item_id = req.params.item_id;
-        let discriminator = req.params.discriminator;
-        await catalog.updateItem(req, item_id, discriminator);
-        // res.redirect('/catalog');
-        res.redirect('/catalog/view/'+discriminator+'/'+item_id);
-    } catch (err) {
-        console.error(err);
-        res.render('error', { error: err });
+    if (currentUserIsAdmin(req)){
+        try {
+            let item_id = req.params.item_id;
+            let discriminator = req.params.discriminator;
+            await catalog.updateItem(req, item_id, discriminator);
+            // res.redirect('/catalog');
+            res.redirect('/catalog/view/'+discriminator+'/'+item_id);
+        } catch (err) {
+            console.error(err);
+            res.render('error', { error: err });
+        } 
+    } else {
+        res.render('index', { title: 'Home', is_logged: req.session.logged, is_admin: req.session.is_admin, errors: [{msg: "You are not an admin!"}]});
     }
 });
 
@@ -161,14 +186,22 @@ router.post('/update/:discriminator/:item_id', async (req, res) => {
 // ====================================== //
 // DELETE an ITEM from the database
 router.get('/deleteitem/:item_id', async(req, res) => {
-    try {
-        await catalog.deleteItem(req.params.item_id);
-        res.redirect('/catalog'); //refresh the page with the new changes 
-    } catch (err) {
-        console.error(err);
-        res.send("Error " + err);
+    if (currentUserIsAdmin(req)){
+        try {
+            await catalog.deleteItem(req.params.item_id);
+            res.redirect('/catalog'); //refresh the page with the new changes
+        } catch (err) {
+            console.error(err);
+            res.send("Error " + err);
+        }
+    } else {
+        res.render('index', { title: 'Home', is_logged: req.session.logged, is_admin: req.session.is_admin, errors: [{msg: "You are not an admin!"}]});
     }
 });
 
 //keep the next line at the end of this script
 module.exports = router;
+
+let currentUserIsAdmin = function (req){
+    return !!(typeof req.session.is_admin !== 'undefined' && req.session.is_admin);
+};
